refactor: simplify plugin runtime startup paths
