+++ src/cli/command-path-policy.test.ts
@@ -1,10 +1,30 @@
 import { afterEach, describe, expect, it, vi } from "vitest";
-import type { CliCommandCatalogEntry } from "./command-catalog.js";
+import type { CliCommandCatalogEntry, CliCommandPathPolicy } from "./command-catalog.js";
 import {
   resolveCliCatalogCommandPath,
   resolveCliCommandPathPolicy,
   resolveCliNetworkProxyPolicy,
 } from "./command-path-policy.js";
+
+const DEFAULT_EXPECTED_POLICY: CliCommandPathPolicy = {
+  bypassConfigGuard: false,
+  routeConfigGuard: "never",
+  loadPlugins: "never",
+  pluginRegistry: { scope: "all" },
+  hideBanner: false,
+  ensureCliPath: true,
+  networkProxy: "default",
+};
+
+function expectResolvedPolicy(
+  commandPath: string[],
+  expected: Partial<CliCommandPathPolicy>,
+): void {
+  expect(resolveCliCommandPathPolicy(commandPath)).toEqual({
+    ...DEFAULT_EXPECTED_POLICY,
+    ...expected,
+  });
+}
 
 describe("command-path-policy", () => {
   afterEach(() => {
@@ -13,66 +33,45 @@ describe("command-path-policy", () => {
   });
 
   it("resolves status policy with shared startup semantics", () => {
-    expect(resolveCliCommandPathPolicy(["status"])).toEqual({
-      bypassConfigGuard: false,
+    expectResolvedPolicy(["status"], {
       routeConfigGuard: "when-suppressed",
       loadPlugins: "never",
-      hideBanner: false,
+      pluginRegistry: { scope: "channels", installBundledRuntimeDeps: false },
       ensureCliPath: false,
       networkProxy: "bypass",
     });
   });
 
   it("applies exact overrides after broader channel plugin rules", () => {
-    expect(resolveCliCommandPathPolicy(["channels", "send"])).toEqual({
-      bypassConfigGuard: false,
-      routeConfigGuard: "never",
+    expectResolvedPolicy(["channels", "send"], {
       loadPlugins: "always",
-      hideBanner: false,
-      ensureCliPath: true,
-      networkProxy: "default",
+      pluginRegistry: { scope: "configured-channels", installBundledRuntimeDeps: false },
     });
-    expect(resolveCliCommandPathPolicy(["channels", "add"])).toEqual({
-      bypassConfigGuard: false,
-      routeConfigGuard: "never",
+    expectResolvedPolicy(["channels", "add"], {
       loadPlugins: "never",
-      hideBanner: false,
-      ensureCliPath: true,
+      pluginRegistry: { scope: "configured-channels", installBundledRuntimeDeps: false },
       networkProxy: "bypass",
     });
-    expect(resolveCliCommandPathPolicy(["channels", "status"])).toEqual({
-      bypassConfigGuard: false,
-      routeConfigGuard: "never",
+    expectResolvedPolicy(["channels", "status"], {
       loadPlugins: "never",
-      hideBanner: false,
-      ensureCliPath: true,
+      pluginRegistry: { scope: "configured-channels", installBundledRuntimeDeps: false },
       networkProxy: expect.any(Function),
     });
-    expect(resolveCliCommandPathPolicy(["channels", "list"])).toEqual({
-      bypassConfigGuard: false,
-      routeConfigGuard: "never",
+    expectResolvedPolicy(["channels", "list"], {
       loadPlugins: "never",
-      hideBanner: false,
-      ensureCliPath: true,
+      pluginRegistry: { scope: "configured-channels", installBundledRuntimeDeps: false },
       networkProxy: "bypass",
     });
-    expect(resolveCliCommandPathPolicy(["channels", "logs"])).toEqual({
-      bypassConfigGuard: false,
-      routeConfigGuard: "never",
+    expectResolvedPolicy(["channels", "logs"], {
       loadPlugins: "never",
-      hideBanner: false,
-      ensureCliPath: true,
+      pluginRegistry: { scope: "configured-channels", installBundledRuntimeDeps: false },
       networkProxy: "bypass",
     });
   });
 
   it("keeps config-only agent commands on config-only startup", () => {
-    expect(resolveCliCommandPathPolicy(["agent"])).toEqual({
-      bypassConfigGuard: false,
-      routeConfigGuard: "never",
+    expectResolvedPolicy(["agent"], {
       loadPlugins: expect.any(Function),
-      hideBanner: false,
-      ensureCliPath: true,
       networkProxy: expect.any(Function),
     });
 
@@ -85,49 +84,31 @@ describe("command-path-policy", () => {
       ["agents", "set-identity"],
       ["agents", "delete"],
     ]) {
-      expect(resolveCliCommandPathPolicy(commandPath)).toEqual({
-        bypassConfigGuard: false,
-        routeConfigGuard: "never",
+      expectResolvedPolicy(commandPath, {
         loadPlugins: "never",
-        hideBanner: false,
-        ensureCliPath: true,
         networkProxy: "bypass",
       });
     }
   });
 
   it("resolves mixed startup-only rules", () => {
-    expect(resolveCliCommandPathPolicy(["configure"])).toEqual({
+    expectResolvedPolicy(["configure"], {
       bypassConfigGuard: true,
-      routeConfigGuard: "never",
       loadPlugins: "never",
-      hideBanner: false,
-      ensureCliPath: true,
-      networkProxy: "default",
     });
-    expect(resolveCliCommandPathPolicy(["config", "validate"])).toEqual({
+    expectResolvedPolicy(["config", "validate"], {
       bypassConfigGuard: true,
-      routeConfigGuard: "never",
       loadPlugins: "never",
-      hideBanner: false,
-      ensureCliPath: true,
       networkProxy: "bypass",
     });
-    expect(resolveCliCommandPathPolicy(["gateway", "status"])).toEqual({
-      bypassConfigGuard: false,
+    expectResolvedPolicy(["gateway", "status"], {
       routeConfigGuard: "always",
       loadPlugins: "never",
-      hideBanner: false,
-      ensureCliPath: true,
       networkProxy: "bypass",
     });
-    expect(resolveCliCommandPathPolicy(["plugins", "update"])).toEqual({
-      bypassConfigGuard: false,
-      routeConfigGuard: "never",
+    expectResolvedPolicy(["plugins", "update"], {
       loadPlugins: "never",
       hideBanner: true,
-      ensureCliPath: true,
-      networkProxy: "default",
     });
     for (const commandPath of [
       ["plugins", "install"],
@@ -136,21 +117,13 @@ describe("command-path-policy", () => {
       ["plugins", "registry"],
       ["plugins", "doctor"],
     ]) {
-      expect(resolveCliCommandPathPolicy(commandPath)).toEqual({
-        bypassConfigGuard: false,
-        routeConfigGuard: "never",
+      expectResolvedPolicy(commandPath, {
         loadPlugins: "never",
-        hideBanner: false,
-        ensureCliPath: true,
-        networkProxy: "default",
       });
     }
-    expect(resolveCliCommandPathPolicy(["cron", "list"])).toEqual({
+    expectResolvedPolicy(["cron", "list"], {
       bypassConfigGuard: true,
-      routeConfigGuard: "never",
       loadPlugins: "never",
-      hideBanner: false,
-      ensureCliPath: true,
       networkProxy: "bypass",
     });
   });
