+++ src/cli/command-path-policy.ts
@@ -12,6 +12,7 @@ const DEFAULT_CLI_COMMAND_PATH_POLICY: CliCommandPathPolicy = {
   bypassConfigGuard: false,
   routeConfigGuard: "never",
   loadPlugins: "never",
+  pluginRegistry: { scope: "all" },
   hideBanner: false,
   ensureCliPath: true,
   networkProxy: "default",
