+++ src/cli/plugin-registry-loader.test.ts
@@ -9,12 +9,10 @@ vi.mock("./plugin-registry.js", () => ({
 describe("plugin-registry-loader", () => {
   let originalForceStderr: boolean;
   let ensureCliPluginRegistryLoaded: typeof import("./plugin-registry-loader.js").ensureCliPluginRegistryLoaded;
-  let resolvePluginRegistryLoadPolicyForCommandPath: typeof import("./plugin-registry-loader.js").resolvePluginRegistryLoadPolicyForCommandPath;
   let loggingState: typeof import("../logging/state.js").loggingState;
 
   beforeAll(async () => {
-    ({ ensureCliPluginRegistryLoaded, resolvePluginRegistryLoadPolicyForCommandPath } =
-      await import("./plugin-registry-loader.js"));
+    ({ ensureCliPluginRegistryLoaded } = await import("./plugin-registry-loader.js"));
     ({ loggingState } = await import("../logging/state.js"));
   });
 
@@ -86,24 +84,6 @@ describe("plugin-registry-loader", () => {
     expect(ensurePluginRegistryLoadedMock).toHaveBeenCalledWith({
       scope: "configured-channels",
       installBundledRuntimeDeps: false,
-    });
-  });
-
-  it("maps command paths to plugin registry load policy", () => {
-    expect(resolvePluginRegistryLoadPolicyForCommandPath(["status"])).toEqual({
-      scope: "channels",
-      installBundledRuntimeDeps: false,
-    });
-    expect(resolvePluginRegistryLoadPolicyForCommandPath(["health"])).toEqual({
-      scope: "channels",
-      installBundledRuntimeDeps: false,
-    });
-    expect(resolvePluginRegistryLoadPolicyForCommandPath(["channels", "send"])).toEqual({
-      scope: "configured-channels",
-      installBundledRuntimeDeps: false,
-    });
-    expect(resolvePluginRegistryLoadPolicyForCommandPath(["agents"])).toEqual({
-      scope: "all",
     });
   });
 });
