+++ src/commands/channels/list.ts
@@ -114,7 +114,7 @@ export async function channelsListCommand(
   const includeUsage = opts.usage !== false;
 
   const plugins = listReadOnlyChannelPluginsForConfig(cfg, {
-    includeSetupRuntimeFallback: true,
+    includeSetupFallbackPlugins: true,
   });
 
   const authStore = loadAuthProfileStoreWithoutExternalProfiles();
