+++ src/commands/channels/status-config-format.ts
@@ -60,7 +60,7 @@ export async function formatConfigChannelsStatusLines(
   const sourceConfig = opts?.sourceConfig ?? cfg;
   const plugins = listReadOnlyChannelPluginsForConfig(cfg, {
     activationSourceConfig: sourceConfig,
-    includeSetupRuntimeFallback: false,
+    includeSetupFallbackPlugins: false,
   });
   for (const plugin of plugins) {
     const accountIds = plugin.config.listAccountIds(cfg);
