+++ src/commands/status-runtime-shared.ts
@@ -30,7 +30,7 @@ export async function resolveStatusSecurityAudit(params: {
   const { runSecurityAudit } = await loadSecurityAuditModule();
   const readOnlyPlugins = resolveReadOnlyChannelPluginsForConfig(params.config, {
     activationSourceConfig: params.sourceConfig,
-    includeSetupRuntimeFallback: false,
+    includeSetupFallbackPlugins: false,
   });
   return await runSecurityAudit({
     config: params.config,
