+++ src/infra/channel-summary.ts
@@ -56,7 +56,7 @@ async function listChannelSummaryPlugins(params: {
   const { listReadOnlyChannelPluginsForConfig } = await import("../channels/plugins/read-only.js");
   return listReadOnlyChannelPluginsForConfig(params.cfg, {
     activationSourceConfig: params.sourceConfig,
-    includeSetupRuntimeFallback: false,
+    includeSetupFallbackPlugins: false,
   });
 }
